Revert "rename `cerboshub` -> `hub` for consistency"
This reverts commit 425634bf0d9df725d3fa07242ccf5a07307adb5d.
