+++ docs/modules/configuration/partials/fullconfiguration.adoc
@@ -11,6 +11,13 @@ audit:
   enabled: false # Enabled defines whether audit logging is enabled.
   excludeMetadataKeys: ['authorization'] # ExcludeMetadataKeys defines which gRPC request metadata keys should be excluded from the audit logs. Takes precedence over includeMetadataKeys.
   includeMetadataKeys: ['content-type'] # IncludeMetadataKeys defines which gRPC request metadata keys should be included in the audit logs.
+  cerboshub:
+    ingest: 
+      flushTimeout: 5s # FlushTimeout defines the max allowable timeout for each Ingest request.
+      maxBatchSize: 32 # MaxBatchSize defines the max number of log entries to send in each Ingest request.
+      minFlushInterval: 10s # MinFlushInterval is the minimal duration between Ingest requests.
+      numGoRoutines: 8 # NumGoRoutines defines the max number of goroutines used when streaming log entries from the local DB.
+    local: see audit.local section # Local is configuration for the local Audit backend. See 'audit.local' section for documentation
   file:
     additionalPaths: [stdout] # AdditionalPaths to mirror the log output. Has performance implications. Use with caution.
     logRotation: # LogRotation settings (optional).
@@ -18,13 +25,6 @@ audit:
       maxFileCount: 10 # MaxFileCount sets the maximum number of files to retain.
       maxFileSizeMB: 100 # MaxFileSizeMB sets the maximum size of individual log files in megabytes.
     path: /path/to/file.log # Required. Path to the log file to use as output. The special values stdout and stderr can be used to write to stdout or stderr respectively.
-  hub:
-    ingest: 
-      flushTimeout: 5s # FlushTimeout defines the max allowable timeout for each Ingest request.
-      maxBatchSize: 32 # MaxBatchSize defines the max number of log entries to send in each Ingest request.
-      minFlushInterval: 10s # MinFlushInterval is the minimal duration between Ingest requests.
-      numGoRoutines: 8 # NumGoRoutines defines the max number of goroutines used when streaming log entries from the local DB.
-    local: see audit.local section # Local is configuration for the local Audit backend. See 'audit.local' section for documentation
   kafka:
     ack: all # Ack mode for producing messages. Valid values are "none", "leader" or "all" (default). Idempotency is disabled when mode is not "all".
     authentication: # Authentication
