+++ internal/test/mocks/IngestSyncer.go
@@ -9,7 +9,6 @@ import (
 	context "context"
 
 	logsv1 "github.com/cerbos/cloud-api/genpb/cerbos/cloud/logs/v1"
-
 	mock "github.com/stretchr/testify/mock"
 )
